L_2

+++ src/main/java/task2/Application.java
@@ -56,7 +56,7 @@ class SquareStickAdapter extends RoundStick {
     public int getRadius() {
         // Вычислить половину диагонали квадратной палки по
         // теореме Пифагора.
-        return stick.getWidth() * Math.sqrt(2) / 2;
+        return (int) (stick.getWidth() * Math.sqrt(2) / 2);
     }
 }
